writing tiles up to 1x1 tile.

+++ src/tasks/CreateTileTask.h
@@ -152,8 +152,8 @@ private:
 
         //TODO check this kind of conversion throughout
         //in particular : when we have size_t as inputs, how do we ensure there is no overflow
-        auto downsampleWidth = static_cast<size_t>(ceil( (size_t)width / 2));
-        auto downsampleHeight = static_cast<size_t>(ceil( (size_t)height / 2));
+        auto downsampleWidth = static_cast<size_t>(ceil( (double)width / 2));
+        auto downsampleHeight = static_cast<size_t>(ceil( (double)height / 2));
 
         T* downsampleData = new T[ downsampleWidth * downsampleHeight ]();
